+++ _temp.py
@@ -1,36 +1,23 @@
-
-from pyDes import des, CBC, PAD_PKCS5
+from Crypto.PublicKey import RSA
+from Crypto.Cipher import PKCS1_OAEP
 import binascii
- 
-# Secret key 
-KEY='abcdefgh'
-def des_encrypt(s):
-    """
-         DES encryption
-         :param s: raw string
-         :return: Encrypted string, hexadecimal
-    """
-    secret_key = KEY
-    iv = secret_key
-    k = des(secret_key, CBC, iv, pad=None, padmode=PAD_PKCS5)
-    en = k.encrypt(s, padmode=PAD_PKCS5)
-    return binascii.b2a_hex(en)
- 
- 
-def des_descrypt(s):
-    """
-         DES decryption
-         :param s: encrypted string, hexadecimal
-         :return: the decrypted string
-    """
-    secret_key = KEY
-    iv = secret_key
-    k = des(secret_key, CBC, iv, pad=None, padmode=PAD_PKCS5)
-    de = k.decrypt(binascii.a2b_hex(s), padmode=PAD_PKCS5)
-    return de
- 
-s='Have a nice day'
-enc=des_encrypt(s)
-print(enc)
-des=des_descrypt(enc)
-print(des)
+
+keyPair = RSA.generate(3072)
+
+pubKey = keyPair.publickey()
+print(f"Public key:  (n={hex(pubKey.n)}, e={hex(pubKey.e)})")
+pubKeyPEM = pubKey.exportKey()
+print(pubKeyPEM.decode('ascii'))
+
+print(f"Private key: (n={hex(pubKey.n)}, d={hex(keyPair.d)})")
+privKeyPEM = keyPair.exportKey()
+print(privKeyPEM.decode('ascii'))
+
+msg = b'A message for encryption'
+encryptor = PKCS1_OAEP.new(pubKey)
+encrypted = encryptor.encrypt(msg)
+print("Encrypted:", binascii.hexlify(encrypted))
+
+decryptor = PKCS1_OAEP.new(keyPair)
+decrypted = decryptor.decrypt(encrypted)
+print('Decrypted:', decrypted)
